+++ core/components/googlesitemap09/docs/changelog.md
@@ -4,3 +4,4 @@
 # GoogleSitemap0.9 1.1.0
 - Added Image Node per resource
 - friendly urls must be turned ON
+- added check for deleted == 1